Course crud Icon fiX

+++ core/views.py
@@ -10,8 +10,7 @@ from django.contrib.auth.decorators import login_required
 
 @login_required
 def Index(request):
-    note = Note.objects.all().order_by('-id')
-    courses = Course.objects.all() # Retrieve Data from the Database
+    courses = Course.objects.filter(author=request.user).order_by('-created') # Retrieve Data from the Database
     context = {'courses':courses} # Pass retrieved data to the frontend
 
     return render(request, 'core/admin/index.html', context)
@@ -21,15 +20,15 @@ def Index(request):
 
 
 # Course  C.R.U.D.
-
+@login_required
 def CourseDetail(request, pk):
     course = Course.objects.get(id=pk)
-    if course:
+    if course.author == request.user:
         notes = Note.objects.filter(course=course)
     context = {'course':course, 'notes':notes}
     return render(request, 'core/note/notes.html', context)
 
-
+@login_required
 def CreateCourse(request):
     if request.method == 'POST':
         form = CreateCourseForm(request.POST)
@@ -49,23 +48,26 @@ def CreateCourse(request):
     context = {'form':form}
     return render(request, 'core/course/create_course.html', context)
 
+@login_required
 def UpdateCourse(request, pk):
     course = Course.objects.filter(id=pk)
-    if request.method == 'POST':
-        form = CreateCourseForm(instance=course)
-        if form.is_valid():
-            form.save()
+    if course.author == request.user:
+        if request.method == 'POST':
+            form = CreateCourseForm(instance=course)
+            if form.is_valid():
+                form.save()
+            else:
+                messages.success(request, 'Invalid Form')
         else:
-            messages.success(request, 'Invalid Form')
-    else:
-        form = CreateCourseForm()
+            form = CreateCourseForm()
     context = {"form":form}
     return render(request, 'core/course/update_course.html', context)
 
 
 # Course notes C.R.U.D.
-def CreateNote(request):
-    course = Course.objects.get(pk=request.session['course_id'])
+@login_required
+def CreateNote(request, pk):
+    course = Course.objects.get(pk=pk, author=request.user)
     if request.method == 'POST':
         form = CreateNoteForm(request.POST)
         if form.is_valid():
@@ -74,7 +76,7 @@ def CreateNote(request):
             var.save()
 
             messages.info(request, 'Note created')
-            return redirect('Index')
+            return redirect('CourseDetail', course.pk)
         else:
             messages.warning(request, 'Sorry something went wrong')
     else:
@@ -82,26 +84,38 @@ def CreateNote(request):
     context = {'form':form}
     return render(request, 'core/note/create.html', context)
 
+@login_required
 def UpdateNote(request, pk):
     note = Note.objects.get(id=pk)
-    form = CreateNoteForm(request.POST or None, instance=note)
+    form = CreateNoteForm(instance=note)
+
+    if note.course.author == request.user:
+        if request.method == 'POST':
+            form = CreateNoteForm(request.POST or None, instance=note)
+            
+            if form.is_valid():
+                form.save()
+                messages.success(request, 'Success....')
+                return redirect('DetailNote', note.pk)
+    
     context = {'form':form}
-    if form.is_valid():
-        form.save()
-        messages.success(request, 'Success....')
-        return redirect('Index')
     return render(request, 'core/note/update.html', context)
 
+@login_required
 def DetailNote(request, pk):
     note = Note.objects.get(id=pk)
-    context = {'note':note}
-    return render(request, 'core/note/detail.html', context)
+    if note.course.author == request.user:
+        context = {'note':note}
+        return render(request, 'core/note/detail.html', context)
 
+@login_required
 def DeleteNote(request, pk):
+
     note = Note.objects.get(id=pk)
-    note.delete()
-    messages.success(request, 'This note has been successfully deleted...')
-    return redirect('Index')
+    if note.course.author == request.user:
+        note.delete()
+        messages.success(request, 'This note has been successfully deleted...')
+        return redirect('Index')
